requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
contourpy
cycler
Flask
fonttools
gitdb
GitPython
idna
itsdangerous
Jinja2
jsonschema
jsonschema-specifications
kiwisolver
MarkupSafe
matplotlib
narwhals
numpy
packaging
pandas
pillow
protobuf
pyarrow
pydeck
pyparsing
python-dateutil
pytz
referencing
requests
rpds-py
six
smmap
streamlit
tenacity
toml
tornado
typing_extensions
tzdata
urllib3
watchdog
Werkzeug
psycopg2-binary
SQLAlchemy
lightrag-hku
google-genai